Clean up linting issues in ssg/xccdf.py
